refactor: useCartProductsStore 폴더명 stores로 변경
- 현재는 하나의 store만 사용하지만 확장성을 고려해 stores로 선언
